The requests are now monitored. The user has access only to his files and reports.

+++ Reports/Controllers/ReportController.cs
@@ -1,7 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using Reports.Authentication;
 using Reports.Entities;
 using Reports.Services;
 using System;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace Reports.Controllers
@@ -9,24 +12,31 @@ namespace Reports.Controllers
     public class ReportController : ApiControllerBase
     {
         private readonly IReportService _reportService;
+        private readonly IUserService _userService;
 
-        public ReportController(IReportService reportService)
+        public ReportController(IReportService reportService, IUserService userService)
         {
             _reportService = reportService;
+            _userService = userService;
         }
 
         [HttpGet]
-        public async Task<IActionResult> Get(int reportId)
+        public async Task<IActionResult> GetById(int reportId)
         {
             try
             {
-                var report = await _reportService.Get(reportId);
+                var report = await _reportService.GetById(reportId);
+                var user = await _userService.GetById(report.UserId);
 
-                return Ok(report);
+                if (GetCurrentUserName() == user.Login)
+                {
+                    return Ok(report); 
+                }
+
+                return BadRequest();
             }
             catch (Exception)
             {
-
                 throw;
             }
         }
@@ -36,13 +46,19 @@ namespace Reports.Controllers
         {
             try
             {
-                var reportId = await _reportService.Create(report);
+                var user = await _userService.GetById(report.UserId);
 
-                return Ok(reportId);
+                if (GetCurrentUserName() == user.Login)
+                {
+                    var reportId = await _reportService.Create(report);
+
+                    return Ok(reportId); 
+                }
+
+                return BadRequest();
             }
             catch (Exception)
             {
-
                 throw;
             }
         }
@@ -52,13 +68,19 @@ namespace Reports.Controllers
         {
             try
             {
-                await _reportService.Update(report);
+                var user = await _userService.GetById(report.UserId);
 
-                return Ok();
+                if (GetCurrentUserName() == user.Login)
+                {
+                    await _reportService.Update(report);
+
+                    return Ok(); 
+                }
+
+                return BadRequest();
             }
             catch (Exception)
             {
-
                 throw;
             }
         }
@@ -68,15 +90,24 @@ namespace Reports.Controllers
         {
             try
             {
-                await _reportService.Delete(reportId);
+                var report = await _reportService.GetById(reportId);
+                var user = await _userService.GetById(report.UserId);
 
-                return Ok();
+                if (GetCurrentUserName() == user.Login)
+                {
+                    await _reportService.Delete(reportId);
+
+                    return Ok(); 
+                }
+
+                return BadRequest();
             }
             catch (Exception)
             {
-
                 throw;
             }
         }
+
+        private string GetCurrentUserName() => User.FindFirstValue(ClaimTypes.Name);
     }
 }
